### 2025-05-07 #8176 @kthyng
@simonbesnard1 Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly.
