+++ paper/refs.bib
@@ -1,11 +1,15 @@
 @article{dubayah2020global,
-  title={The Global Ecosystem Dynamics Investigation: High-resolution laser ranging of the Earth’s forests and topography},
-  author={Dubayah, Ralph and Blair, James Bryan and Goetz, Scott and Fatoyinbo, Lola and Hansen, Matthew and Healey, Sean and Hofton, Michelle and Hurtt, George and Kellner, James and Luthcke, Scott and others},
-  journal={Science of remote sensing},
-  volume={1},
-  pages={100002},
-  year={2020},
-  publisher={Elsevier}
+title = {The Global Ecosystem Dynamics Investigation: High-resolution laser ranging of the Earth’s forests and topography},
+journal = {Science of Remote Sensing},
+volume = {1},
+pages = {100002},
+year = {2020},
+issn = {2666-0172},
+doi = {10.1016/j.srs.2020.100002},
+url = {https://www.sciencedirect.com/science/article/pii/S2666017220300018},
+author = {Ralph Dubayah and James Bryan Blair and Scott Goetz and Lola Fatoyinbo and Matthew Hansen and Sean Healey and Michelle Hofton and George Hurtt and James Kellner and Scott Luthcke and John Armston and Hao Tang and Laura Duncanson and Steven Hancock and Patrick Jantz and Suzanne Marselis and Paul L. Patterson and Wenlu Qi and Carlos Silva},
+keywords = {Lidar, Ecosystem structure, GEDI, Biomass},
+abstract = {Obtaining accurate and widespread measurements of the vertical structure of the Earth’s forests has been a long-sought goal for the ecological community. Such observations are critical for accurately assessing the existing biomass of forests, and how changes in this biomass caused by human activities or variations in climate may impact atmospheric CO2 concentrations. Additionally, the three-dimensional structure of forests is a key component of habitat quality and biodiversity at local to regional scales. The Global Ecosystem Dynamics Investigation (GEDI) was launched to the International Space Station in late 2018 to provide high-quality measurements of forest vertical structure in temperate and tropical forests between 51.6° N & S latitude. The GEDI instrument is a geodetic-class laser altimeter/waveform lidar comprised of 3 lasers that produce 8 transects of structural information. Over its two-year nominal lifetime GEDI is anticipated to provide over 10 billion waveforms at a footprint resolution of 25 ​m. These data will be used to derive a variety of footprint and gridded products, including canopy height, canopy foliar profiles, Leaf Area Index (LAI), sub-canopy topography and biomass. Additionally, data from GEDI are used to demonstrate the efficacy of its measurements for prognostic ecosystem modeling, habit and biodiversity studies, and for fusion using radar and other remote sensing instruments. GEDI science and technology are unique: no other space-based mission has been created that is specifically optimized for retrieving vegetation vertical structure. As such, GEDI promises to advance our understanding of the importance of canopy vertical variations within an ecological paradigm based on structure, composition and function.}
 }
 
 @misc{dubayah2021gedi_l2a,
@@ -156,6 +160,7 @@
       eprint={2406.01076},
       archivePrefix={arXiv},
       primaryClass={cs.CV},
+      doi={10.48550/arXiv.2406.01076,}
       url={https://arxiv.org/abs/2406.01076}, 
 }
 
@@ -166,7 +171,7 @@ volume = {308},
 pages = {114174},
 year = {2024},
 issn = {0034-4257},
-doi = {https://doi.org/10.1016/j.rse.2024.114174},
+doi = {10.1016/j.rse.2024.114174},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425724001925},
[…]
