### 2025-05-08 #8176 @simonbesnard1
Thanks, @kthyng, for considering my paper for JOSS.

I have fixed the missing doi in the refs.bib file and shortened the text to comply with the journal requirements. Later today, I will push those changes to the main branch.

Please find below a list of people who could review the paper:

- Loic Dutrieux (European Commission, Joint Research Centre)
- Jan Pauls (University of Munster)
- Chuck Daniels (Seed)
- Ralph Dubayah (University of Maryland)
- Konstantin Ntokas (Brockmann Consult GmbH )

Let me know if you need their specific contacts.
