+++ paper/refs.bib
@@ -190,3 +190,11 @@ abstract = {More of the Amazon rainforest is disturbed each year than completely
   url = {https://www.nature.com/articles/s41586-024-07629-0},
   publisher={Nature Publishing Group UK London}
 }
+
+@Manual{tiledb2025,
+  title = {tiledb: Modern Database Engine for Complex Data Based on Multi-Dimensional Arrays},
+  author = {{TileDB, Inc.}},
+  year = {2025},
+  note = {Pyhton package version 0.33.6},
+  url = {https://github.com/TileDB-Inc/TileDB-Py},
+}
