### 2025-05-08 70adcae
fix refs.bib bug
