+++ paper/refs.bib
@@ -160,7 +160,7 @@ abstract = {Obtaining accurate and widespread measurements of the vertical struc
       eprint={2406.01076},
       archivePrefix={arXiv},
       primaryClass={cs.CV},
-      doi={10.48550/arXiv.2406.01076,}
+      doi={10.48550/arXiv.2406.01076},
       url={https://arxiv.org/abs/2406.01076}, 
 }
 
