### 2025-07-02 #8176 @simonbesnard1
@kanishkan91, would you happen to have any updates on the review of our paper?

Thanks for letting me know.
Simon.
