### 2025-07-14 #8176 @kanishkan91
@maruti-iitm , @sinanozaydin - Would eaither/both of you be interested in reviewing this manuscript for JOSS, given your previous reviews, interests?
