### 2025-07-14 #8176 @sinanozaydin
> [@maruti-iitm](https://github.com/maruti-iitm) , [@sinanozaydin](https://github.com/sinanozaydin) - Would eaither/both of you be interested in reviewing this manuscript for JOSS, given your previous reviews, interests?

Hello @kanishkan91 ! Unfortunately, this is quite out of my expertise.

Cheers,

Sinan
