### 2025-07-15 #8176 @kanishkan91
@shubhamjain15 I know you just did a review a while back, but are you interested in reviewing this paper for JOSS? Seemed like a good fit.
