### 2025-07-15 #8593 @editorialbot
[REVIEW]: gediDB: A toolbox for processing and providing Global Ecosystem Dynamics Investigation (GEDI) L2A-B and L4A-C data
**Submitting author:** @simonbesnard1 (<a href="http://orcid.org/0000-0002-1137-103X">Simon Besnard</a>)
**Repository:** https://github.com/simonbesnard1/gedidb
**Branch with paper.md** (empty if default branch):
**Version:** 2025.9.24
