### 2025-09-01 simonbesnard1/gedidb#33 @shubhamjain15
DOC: Windows install - Earthdata authentication fails due to missing wget external dependency
### Issue with current documentation:

On Windows, `EarthDataAuthenticator.authenticate()` fails because the package calls the `wget` CLI, which is not installed by default on many Windows/Python environments.

Please document `wget` as an external dependency in the user guide.

(From JOSS review: openjournals/joss-reviews#8593)
