+++ gedidb/core/gediprocessor.py
@@ -20,7 +20,6 @@ import yaml
 from dask.distributed import Client
 
 from gedidb.core.gedidatabase import GEDIDatabase
-<<<<<<< HEAD
 from gedidb.core.gedigranule import GEDIGranule
 from gedidb.downloader.authentication import EarthDataAuthenticator
 from gedidb.downloader.data_downloader import CMRDataDownloader, H5FileDownloader
@@ -261,49 +260,6 @@ class GEDIProcessor:
             "No parallel engine provided. Defaulting to single-threaded execution."
         )
         return concurrent.futures.ThreadPoolExecutor(max_workers=1)
-=======
-from gedidb.utils.geospatial_tools import ShapeProcessor
-
-
-logger = logging.getLogger(__name__)
-
-def log_execution(start_message=None, end_message=None):
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            logger.info(start_message or f"Executing {func.__name__}...")
-            result = func(*args, **kwargs)
-            logger.info(end_message or f"Finished {func.__name__}...")
-            return result
-        return wrapper
-    return decorator
-    
-class GEDIGranuleProcessor(GEDIDatabase):
-    
-    def __init__(self, config_files: dict):
-        self.load_all_configs(config_files)
-        super().__init__(
-            self.data_info['region_of_interest'], 
-            self.data_info['start_date'], 
-            self.data_info['end_date']
-        )
-        self.setup_paths_and_dates()
-    
-    def load_all_configs(self, config_files):
-        """Load all configuration files."""
-        self.data_info = self.load_config_file(config_files['data_info'])
-        self.database_schema = self.load_config_file(config_files['database_schema'])
-
-    def setup_paths_and_dates(self):
-        """Set up paths and dates based on the configuration."""
-        self.download_path = self.ensure_directory(self.data_info['download_path'])
-        self.parquet_path = self.ensure_directory(self.data_info['parquet_path'])
-        self.db_path = self.data_info['database_url']
-        initial_geom = gpd.read_file(self.data_info['region_of_interest'])
-        self.geom = ShapeProcessor(initial_geom).check_and_format(simplify=True)        
-        self.start_date = datetime.strptime(self.data_info['start_date'], '%Y-%m-%d')
-        self.end_date = datetime.strptime(self.data_info['end_date'], '%Y-%m-%d')
->>>>>>> progress config restructuring
 
     @staticmethod
     def _ensure_directory(path: str) -> str:
@@ -316,7 +272,6 @@ class GEDIGranuleProcessor(GEDIDatabase):
         """Load a YAML configuration file."""
         with open(file_path, "r") as file:
             return yaml.safe_load(file)
-<<<<<<< HEAD
 
     def compute(self, consolidate: bool = True, consolidation_type: str = "spatial"):
         """
@@ -329,81 +284,6 @@ class GEDIGranuleProcessor(GEDIDatabase):
         consolidation_type : str, default='spatial'
             Type of consolidation to perform ('default' or 'spatial').
         """
-=======
-                
-    @log_execution(start_message = "Starting computation process...", end_message='Data processing completed!')
-    def compute(self):
-        cmr_data = self.download_cmr_data().download()
-        spark = self.create_spark_session()
-    
-        name_url = cmr_data[
-            ["id", "name", "url", "product"]
-        ].to_records(index=False)
-    
-        urls = spark.sparkContext.parallelize(name_url)
-        downloader = H5FileDownloader(self.download_path)
-    
-        mapped_urls = urls.map(lambda x: downloader.download(x[0], x[2], GediProduct(x[3]))).groupByKey()
[…]
