+++ gedidb/processor/granule/l2a_granule.py
@@ -1,21 +1,0 @@
-from gedidb.processor.granule.granule import Granule
-from gedidb.processor.beam.l2a_beam import L2ABeam
-from gedidb.processor.beam.beam import Beam
-
-
-class L2AGranule(Granule):
-
-    def __init__(self, file_path, quality_flag, field_mapping, geom):
-        super().__init__(file_path)
-        
-        self.quality_flag = quality_flag
-        self.field_mapping = field_mapping
-        self.geom = geom
-        
-        print(self.version)
-    def _beam_from_name(self, beam: str) -> Beam:
-
-        if beam not in self.beam_names:
-            raise ValueError(f"Beam name must be one of {self.beam_names}")
-        return L2ABeam(self, beam, self.quality_flag, self.field_mapping, self.geom)
-
