### 2025-09-09 simonbesnard1/gedidb#39 @jhkennedy
JOSS review summary issue
👋 I'm wrapping up my review for JOSS (https://github.com/openjournals/joss-reviews/issues/8593), and first I want to say, this is very nice work!

There's a few issues/PRs I'm going to open. I'm going to use this issue as a summary issue, and will edit the description as I go. When I've got everything open and recorded here, I'll comment below stating my initial review is done.

---

Docs + Community files:

- [x] #40
- [x] #41
- [x] #42
- [x] #43
- [x] #44
- [x] #46
- [x] #47
- [x] #50

- [x] #45
- [x] #48

Paper
- [x] #49
