+++ .github/ISSUE_TEMPLATE/documentation.yml
@@ -1,5 +1,5 @@
 name: Documentation
-description: Report an issue related to the NumPy documentation.
+description: Report an issue related to the gedidb documentation.
 title: "DOC: <Please write a comprehensive title after the 'DOC: ' prefix>"
 labels: [04 - Documentation]
 
