### 2025-09-09 simonbesnard1/gedidb#42 @jhkennedy
DOC: Change Quick Start order in menu for ease of use
### Issue with current documentation:

I have to say, overall the documentation is really well done -- it looks very professional and is generally easy to read.

That said, the location of the Quick Start tripped me up a couple of time.

When you first open the Docs link, you're prompted with "Getting Started" as the first box, but when you click on that box, you skip over installation. And if you navigate to "next" instead of Quick Start you end up on installation:

If you do click the quick start, it jumps right into usage, and has skipped over installation in the main menu:

I think it'd be much more intuitive if either:
1. the quick start included the preferred installation `pip install gedidb` or `pip install gedidb[full]` at the top with a link back to detailed installation steps, or
2. In the users guide put installation as the first box

I prefer (1), personally.

---

This Issue is part of https://github.com/openjournals/joss-reviews/issues/8593; see the meta-tracking issue here: #39

### Idea or request for content:

_No response_
