+++ doc/user/tiledb_database.rst
@@ -253,7 +253,7 @@ Here are some example use cases:
 Resources
 ---------
 - `TileDB Documentation <https://tiledb.com/docs>`_
-- `gediDB GitLab Repository <https://github.com/simonbesnard1/gedidb>`_
+- `gediDB GitHub Repository <https://github.com/simonbesnard1/gedidb>`_
 - `GEDI Data Products Overview <https://gedi.umd.edu>`_
 
    
