### 2025-09-09 simonbesnard1/gedidb#44 @jhkennedy
DOC: Minor reorganization suggestions
### Issue with current documentation:

_No response_

### Idea or request for content:

There's a couple pages I think would fit better in a different sections:
1. "Contributing to gediDB" in "Community"
2. "API Reference" in "Advanced Usage" or maybe "Fundamentals"
3. "Frequently Asked Questions (FAQ)"  in "Community"

_Note: These especially are just suggestions from coming to the package as a first time user, and it's totally fine to reject/disagree with them 😄 ._

---

This Issue is part of https://github.com/openjournals/joss-reviews/issues/8593; see the meta-tracking issue here: #39
