### 2025-09-09 simonbesnard1/gedidb#45 @jhkennedy
BUG: netrc appended when credentials already exist
### Describe the issue:

Following the [NASA authentication](https://gedidb.readthedocs.io/en/latest/user/fundamentals.authenticate.html) directions in the docs:
```python
import gedidb as gdb

authentificator = gdb.EarthDataAuthenticator()
authentificator.authenticate()
```

If you already have a `~/.netrc` file setup, but don't have the `~/.cookies` file, gedidb will prompt you for your username and password and then append it to the file again.

For example, before running the authenticator my `.netrc` file looked like:
```
machine urs.earthdata.nasa.gov login [REDACTED] password [REDACTED]

```

Running those commands resulted in:
```
>>> import gedidb as gdb
>>> authentificator = gdb.EarthDataAuthenticator()
>>> authentificator.authenticate()
2025-09-09 13:32:01,149 - WARNING - Missing cookie file at /home/jhkennedy/.cookies.
2025-09-09 13:32:01,149 - INFO - EarthData authentication setup incomplete; starting setup.
2025-09-09 13:32:01,149 - INFO - Prompting user for credentials.
Please enter your Earthdata Login username: >? [REDACTED]
Warning: Password input may be echoed.
Please enter your Earthdata Login password: >? [REDACTED]
2025-09-09 13:32:27,499 - INFO - Credentials added to .netrc file.
2025-09-09 13:32:27,499 - INFO - Attempting to fetch Earthdata cookies and save to /home/jhkennedy/.cookies
2025-09-09 13:32:28,163 - INFO - Earthdata cookies successfully fetched and saved to /home/jhkennedy/.cookies.
```

And after, it looked like:
```
machine urs.earthdata.nasa.gov login [REDACTED] password [REDACTED]

machine urs.earthdata.nasa.gov login [REDACTED] password [REDACTED]
```

Ideally, the authenticator wouldn't append to my netrc file if `machine urs.earthdata.nasa.gov` wasn't already listed there, and it shouldn't need to prompt in that case either.

### Reproduce the code example:

```python
import gedidb as gdb

authentificator = gdb.EarthDataAuthenticator()
authentificator.authenticate()
```

### Python and gedidb Versions:

Python: 3.15.5
gedidb: 2025.4.26.dev27+g7036217ea

### Context for the issue:

---

This Issue is part of https://github.com/openjournals/joss-reviews/issues/8593; see the meta-tracking issue here: #39
