### 2025-09-10 simonbesnard1/gedidb#47 @jhkennedy
DOC: Quck Overview suggestions
### Issue with current documentation:

When trying out gediDB for the first time, the quick Overview wasn't very self contained, and I think could better demonstrate the value of gedidb immediately.

When I walked through it, I:
1. was missing the install instructions (also discussed in #42) and had to hop back to the install section
2. felt the `config_file.yml` and `test.geojson` files were opaque and wasn't sure if I needed to create them or if they would be created for me in that location. I had to hope to the configurations files section and from there, i just downloaded the `data_config.yml` file. Inside that  file and wasn't sure what `local_path` in particular needed to be (a directory that already exists, apparently). I also had to edit `data_dir`.
3. bonked against NASA Earthdata credentials as they were not mentioned on the Quick Overview or on the Configuration Files page and the only hint was this line: `earth_data_dir= '/path/to/earthdata_credential/folder',`. That at least made me search the docs to figure out how NASA auth was supposed to be provided.

Overall, I was able to work through those issues from the documentation, which is great, but it wasn't any quicker than just going through the documentation in order.

I also didn't come away with a clear motivation for me to use gedidb -- the quick overview looks remarkably similar to downloading and processing the data myself. I think generally the power here is that (correct me if I'm wrong) once you have the data in the gediDB, querying and using the data is faster and easier, and you don't have to think about granules or auth or the HDF5 files/structure anymore.

With that in mind, for the quick overview, I think it'd be way more powerful to focus on demonstrating the ease of access part. That is:
1. add minimal, preferred install instructions
2. drop the processing section from the quick overview
3. switch the Querying section to pulling data from the Global Land Monitoring group's TileDB at `https://s3.gfz-potsdam.de`.
4. (maybe?) add a plot or analysis of the data pulled in (3)

That would get you up and running directly and you could direct users to set the documentation pages for setting up their own database, and importantly why they would want to do so.

If I run through the whole quick overview, and time it, I can download all the data locally (to my laptop in AK), process the data, and then query it in about 13 minutes. If instead I just query the data from postdam with everything else the same as in the quick overview, which is a pretty slow network hop, I get back 4x the data in 5 minutes.

I think the ease and speed you can get data into users hands is the compelling story to tell in the quick overview, and then you can discuss why you wouldn't just use the postdam data and would want to set up your own database to query and link out to the main docs.

### Idea or request for content:

_No response_

---

This Issue is part of https://github.com/openjournals/joss-reviews/issues/8593; see the meta-tracking issue here: #39
