### 2025-09-10 simonbesnard1/gedidb#49 @jhkennedy
JOSS Paper: Review comments
### What is your issue?

Overall the paper is well written and and well structured and was a pleasure to read.

I only have a few comments:
1. I like figure 2 and the "handwriting" font for the titles/headers, but the lists inside the grey boxes are hard for me to read. Maybe using a lighter grey or just an outline would alleviate it. Note: the [same figure](https://gedidb.readthedocs.io/en/latest/user/why-gedidb.html) is in the docs.
2. It looks like the "Statement of Need" covers "the state of the feild, but I think it missed a lot of similar work has been done in the ICESat-2 community. How this compares with [SlideRule](https://joss.theoj.org/papers/10.21105/joss.04982) would be fitting here, if only to motivate this is a common problem with NASA LiDAR data.
3. The "Performance benchmarks" section doesn't really give me enough information to decide if the query time is good or not, and I would typically expect a comparison to doing it a different way. How long, for example, would it take to do these queries with the GEDI Subsetter via MAAP?
4. In the "Example use cases" section, the illustrative use case is all written in past-tense, without an indication of who was doing it, and particularly, who was able to reproduce it. Nor does it really describe how gediDB does things better than, say, using MAAP.
6. The "Future developments" could either use some fleshing out (who benifits from them and why?) or at least linking our to issues/projects that describe them more. For example, when I read "Support for direct HDF5 access from AWS S3" I expect this to be directly accessing the NASA granules as part of the GEDIProcessor step, but there is a similarly named issue (#15) that is just looking for already downloaded granules before downloading---are they the same?
