### 2025-09-10 simonbesnard1/gedidb#39 @jhkennedy
I have finished my review!

For my closing comments and overall summary, I'd like to say the package will very well done, and the docs in particular look very nice. All the PRs I opened are minor to trivial, and most of the issues I opened are minor Docs enhancements. 👏

I do think spending a little more time on the quick overview section (#47) would be particularly beneficial to new users, but I was able to work through it fine.

The two functionality issues I opened are a largely inconsequential bug (appending already existing info to the netrc; #4) and a minor enhancement.

None of the above is blocking publication in my mind for publication, but would be good to merge/address for users.

There are however some minor things that the paper would benefit from (#49) that should at least be responded to if not addressed before publication, in my opinion.
