### 2025-09-10 simonbesnard1/gedidb#50 @jhkennedy
DOC: Installation docs suggestions
### Issue with current documentation:

I think the installation docs could be substantially improved.

Particularly:
1. I don't find the "Dependencies" section to be of much value, and has a risk of going out of date. The `pyproject.toml` is the canonical place to record the minimum version of things, and notably, the minimum version are not listed there. These "minimum versions" appear to be the versions strictly pinned in the `environment.yml` file instead. The `proproject.toml` and "minimum versions" listed in the docs should be there to enforce known incompatibilities, not to state a known working version which is what a lock file is for. I would reduce this to just a text description of what dependencies are needed, especially since it takes up so much space (the actual install commands aren't even on my screen without scrolling down).
2.  Optional dependency and external dependency sections are good! 👍
3.  Install instructions are good, though `pip` should be prefixed with `python -m pip` now as a best practice
4.  The instructions don't describe setting up a development version at all, which is needed for the Testing section. I think the `Development Version` would be better covered instead by describing how to setup a development environment locally and running the tests.
5. The tests section would better if it told you to install the `dev` target (`python -m pip install .[dev]`) which is in the `pyproject.toml`, but it gives this warning:
   ```
   WARNING: gedidb 2025.4.26.dev43+gb94583811 does not provide the extra 'complete'
   ```
   because there is not `complete` target which it tries to install.

### Idea or request for content:

_No response_
