### 2025-09-10 #8593 @jhkennedy
@simonbesnard1 @kanishkan91 I have finished my review! Sorry for the delay.

Overall,  I'd like to say the package will very well done, and the docs in particular look very nice.

I had a bit of feedback as I went through my review, so I opened summary/tracking issue (https://github.com/simonbesnard1/gedidb/issues/39) wich contains 7 child issues and 4 child PRs.

Over all, they are very minor, but I do think spending a little more time on the quick overview section (https://github.com/simonbesnard1/gedidb/issues/47) and the installation section (https://github.com/simonbesnard1/gedidb/issues/50) of the docs would be particularly beneficial to new users even though I was able to work through both fine.

None of the above is blocking in my mind for publication, but would be good to merge/address for users.

There are however some minor things that the paper would benefit from (https://github.com/simonbesnard1/gedidb/issues/49) that should at least be responded to if not addressed before publication, in my opinion.

@simonbesnard1 I'm happy to discuss any of the issues/PRs raised further, and help address them. Just let me know!

@kanishkan91 I'd recommend accept with Minor Revisions, but happy to defer to you since @shubhamjain15 recommends accepting with no revisions.
