### 2025-09-10 9e1188b
Update contributing.rst
