+++ paper/refs.bib
@@ -198,3 +198,5 @@ abstract = {More of the Amazon rainforest is disturbed each year than completely
   note = {Pyhton package version 0.33.6},
   url = {https://github.com/TileDB-Inc/TileDB-Py},
 }
+
+@article{Shean2023, doi = {10.21105/joss.04982}, url = {https://doi.org/10.21105/joss.04982}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {81}, pages = {4982}, author = {Shean, David and Swinski, J.p. and Smith, Ben and Sutterley, Tyler and Henderson, Scott and Ugarte, Carlos and Lidwa, Eric and Neumann, Thomas}, title = {SlideRule: Enabling rapid, scalable, open science for the NASA ICESat-2 mission and beyond}, journal = {Journal of Open Source Software} }
