### 2025-09-10 77dd371
add Amelia phd reference
