+++ paper/refs.bib
@@ -200,3 +200,15 @@ abstract = {More of the Amazon rainforest is disturbed each year than completely
 }
 
 @article{Shean2023, doi = {10.21105/joss.04982}, url = {https://doi.org/10.21105/joss.04982}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {81}, pages = {4982}, author = {Shean, David and Swinski, J.p. and Smith, Ben and Sutterley, Tyler and Henderson, Scott and Ugarte, Carlos and Lidwa, Eric and Neumann, Thomas}, title = {SlideRule: Enabling rapid, scalable, open science for the NASA ICESat-2 mission and beyond}, journal = {Journal of Open Source Software} }
+
+@article{holcomb_measuring_2025,
+	title = {Measuring tropical forest disturbance and regrowth with spaceborne lidar},
+	url = {https://www.repository.cam.ac.uk/handle/1810/389269},
+	abstract = {The world’s tropical forests are vulnerable, facing threats of deforestation and anthropogenic disturbances, but they are also the site of enormous potential for forest growth. While these changes in tropical forest frontiers are widely agreed to have major implications for global carbon fluxes and biodiversity, it is difficult to quantify their impact using traditional optical and radar remote sensing. As a result, there is wide scientific uncertainty about the scale and trajectory of tropical forest loss and gain, and forest degradation is rarely incorporated into international conservation commitments or land use change reporting. The Global Ecosystem Dynamics Investigation (GEDI), a spaceborne lidar sensor, offers the opportunity to measure forest structure and carbon stocks at global scale, and is sensitive enough to capture subtle forest variation even under dense tropical canopies. However, it is computationally challenging to use GEDI directly to measure change, because of the instrument’s sparse sampling pattern, geolocation uncertainty, and unwieldy data size – billions of measurements comprising hundreds of terabytes of raw data. Indeed, early publications raised explicit doubts that GEDI would be suitable for studying forest disturbance. In this PhD, I develop advanced computational techniques to overcome the limitations of GEDI and make use of these techniques to answer ecological questions about the impact of regrowth and disturbance in tropical forests. In doing so, I argue that GEDI can be used to monitor tropical forest changes – sometimes, in unexpected ways that exceed the limitations assumed by prior research. This finding has important implications for our scientific understanding of carbon fluxes and biodiversity impacts in disturbed tropical forests, for research priorities in distributed geospatial computing to support future spaceborne lidar missions, and for international policy goals.},
+	language = {eng},
+	urldate = {2025-09-10},
+	author = {Holcomb, Amelia},
+	month = sep,
+	year = {2025},
+	file = {Full Text PDF:/home/simon/snap/zotero-snap/common/Zotero/storage/8Y4GBE3A/Holcomb - 2025 - Measuring tropical forest disturbance and regrowth with spaceborne lidar.pdf:application/pdf},
+}
