### 2025-09-11 c7cda17
update fig 3
