### 2025-09-11 a67e377
update example use case
