### 2025-09-11 341ac96
update installing.rst file
