+++ doc/user/quick-overview.rst
@@ -23,7 +23,7 @@ This section provides brief examples of using :py:class:`gedidb.GEDIProcessor` a
       python -c "import gedidb; print(gedidb.__version__)"
 
    For system dependencies, optional extras, and troubleshooting,
-   see :ref:`_installing`.
+   see :ref:`installing`.
 
 ---
 
@@ -37,9 +37,9 @@ Start by importing the **gedidb** package:
 Processing GEDI Data
 --------------------
 
-To process GEDI data, specify paths to a ``YAML`` configuration file (`config_file`). See :ref:`fundamentals-setup` for more information on the data configuration files.
+To process GEDI data, specify paths to a ``YAML`` configuration file (`config_file`). See :ref:`fundamentals-setup` for more information on configuration files.
 
-This setup initiates the download, processing, and storage of GEDI data in your database.
+This setup initiates the **download, processing, and storage** of GEDI data into your database.
 
 .. code-block:: python
 
@@ -48,30 +48,29 @@ This setup initiates the download, processing, and storage of GEDI data in your 
     geometry = 'path/to/test.geojson'
 
     # Initialize a parallel engine
-    concurrent_engine= concurrent.futures.ThreadPoolExecutor(max_workers=10)
+    concurrent_engine = concurrent.futures.ThreadPoolExecutor(max_workers=10)
 
     # Initialize the GEDIProcessor and compute
     with gdb.GEDIProcessor(
         config_file=config_file,
         geometry=geometry,
         start_date='2020-01-01',
-        end_date='2020-12-31',   
-        earth_data_dir= '/path/to/earthdata_credential/folder',
-        parallel_engine=concurrent_engine, 
+        end_date='2020-12-31',
+        earth_data_dir='/path/to/earthdata_credential/folder',
+        parallel_engine=concurrent_engine,
     ) as processor:
         processor.compute(consolidate=True)
 
-
 In this example, the :py:class:`gedidb.GEDIProcessor` performs:
 
-- **Downloading** GEDI L2A-B and L4A-C products.
+- **Downloading** GEDI L2A–B and L4A–C products.
 - **Filtering** data by quality.
-- **Storing** the processed data in the tileDB database.
+- **Storing** the processed data as sparse arrays in the TileDB database.
 
 Querying GEDI Data
 ------------------
 
-Once the data is processed and stored, use :py:class:`gedidb.GEDIProvider` to query it. The results can be returned in either **Xarray** or **Pandas** format, providing flexibility for various workflows.
+Once the data is processed and stored, use :py:class:`gedidb.GEDIProvider` to query it. The results can be returned in either **Xarray** or **Pandas** format, providing flexibility for different workflows.
 
 Example query using :py:class:`gedidb.GEDIProvider`:
 
@@ -81,16 +80,18 @@ Example query using :py:class:`gedidb.GEDIProvider`:
     import gedidb as gdb
 
     # Create GEDIProvider instance
-    provider = gdb.GEDIProvider(storage_type='local', 
-                                local_path= "path/to/your/database/")
+    provider = gdb.GEDIProvider(
+        storage_type='local',
+        local_path="path/to/your/database/"
+    )
 
     # Load region of interest
     region_of_interest = gpd.read_file('./data/geojson/BR-Sa1.geojson')
 
     # Define the columns to query and additional parameters
-    vars_selected = ["agbd", 'rh']
-    
-    # Profile the provider's `get_data` function
+    vars_selected = ["agbd", "rh"]
+
+    # Query the database
     gedi_data = provider.get_data(
         variables=vars_selected,
         query_type="bounding_box",
@@ -100,13 +101,12 @@ Example query using :py:class:`gedidb.GEDIProvider`:
[…]
