### 2025-09-12 b638375
update specs Ceph object
