### 2025-09-13 #8593 @kanishkan91
> 👋 [@simonbesnard1](https://github.com/simonbesnard1) - I'm taking a read through the latest manuscript. We need to set up the archive for your new release.
>
> We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So as next steps for you:
>
> * [ ]
>   [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Please ensure that the software archive uses the same license as the license you have posted on GitHub.
>   [ ]
>   [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.
>   [ ]
>   [ ]  Check that software archive exists and title and author list match JOSS paper (or purposefully do not).
>
> Please respond with the 1) version and 2) DOI of the archived version here

@simonbesnard1 Once you get the above done, I can work on recommending acceptance
