+++ pyproject.toml
@@ -28,19 +28,18 @@ classifiers = [
 ]
 requires-python = ">=3.10"
 
-# ↓ Use *minimum compatible* versions; keep pins only in lockfiles/CI
 dependencies = [
-    "dask[array]>=2023.7",
-    "distributed>=2023.7",
-    "geopandas>=0.13",
-    "h5py>=3.8",
-    "numpy>=1.24",
-    "pandas>=2.0",
-    "requests>=2.31",
+    "dask>=2025.9.1",
+    "distributed>=2025.9.1",
+    "geopandas>=1.1.1",
+    "h5py>=3.14.0",
+    "numpy>=2.3.3",
+    "pandas>=2.3.2",
+    "requests>=2.32.5",
     "retry>=0.9.2",
-    "scipy>=1.10",
-    "tiledb>=0.25",
-    "xarray>=2023.7"
+    "scipy>=1.16.2",
+    "tiledb>=0.34.2",
+    "xarray>=2025.9.0"
 ]
 
 [project.urls]
