### 2025-09-24 da4fd17
update dev installation
