### 2025-09-24 5ec2334
Prep JOSS reviewed release: bump version, sync metadata with paper
