+++ CITATION.cff
@@ -6,12 +6,13 @@ authors:
     orcid: "https://orcid.org/0000-0002-1137-103X"
   - family-names: "Dombrowski"
     given-names: "Felix"
+    orcid: "https://orcid.org/0009-0000-9210-3530"
   - family-names: "Holcomb"
     given-names: "Amelia"
     orcid: "https://orcid.org/0000-0001-5081-7201"
-title: "gediDB"
+title: "gediDB: A toolbox for processing and providing Global Ecosystem Dynamics Investigation (GEDI) L2A-B and L4A-C data"
 abstract: "A toolbox for Global Ecosystem Dynamics Investigation (GEDI) L2A-B and L4A-C data."
 license: "EUPL-1.2"
 doi: "https://doi.org/10.5281/zenodo.13885229"
-repository-code: "https://git.gfz-potsdam.de/global-land-monitoring/gedidb"
+repository-code: "https://github.com/simonbesnard1/gedidb"
 
