### 2025-09-24 #8593 @simonbesnard1
@kanishkan91 Below are the information you need:

version of gediDB: 2025.9.24
DOI in Zenodo: https://doi.org/10.5281/zenodo.17191670

A new release was done in PyPi

https://pypi.org/project/gedidb/

Please let me know if you need anything else.
