### 2025-09-24 771c663
update citation
