### 2025-09-25 #8593 @kanishkan91
@simonbesnard1  Looks like there was an error with some references in the pdf (See above). Could you correct those?
