### 2025-09-25 #8593 @kanishkan91
@simonbesnard1 Once you correct the above, I can recommend this for acceptance. I will be reading through the paper for other typos etc. shortly. The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @shubhamjain15 and @jhkennedy  for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
