+++ paper/paper.md
@@ -39,8 +39,8 @@ The Global Ecosystem Dynamics Investigation (GEDI) mission provides spaceborne L
 
 High-volume LiDAR datasets from the Global Ecosystem Dynamics Investigation (GEDI) mission [@dubayah2020global] ([Fig. 1](#fig1)) have become essential for quantifying forest dynamics, estimating biomass, and analysing carbon cycling. The open availability of GEDI’s spaceborne LiDAR data has enabled forest structural analysis at near-global scales. However, practical use remains hindered by the complexity of raw HDF5 granules, the absence of scalable infrastructure for efficient access, and a lack of standardised tools for large-scale spatial and temporal subsetting.
 
-![<a name="fig1"></a>Schematic representation of the GEDI data structure](figs/beam_product_footprint.png)
-*Fig. 1: A schematic representation of the GEDI data structure. Credits: Amelia Holcomb's PhD dissertation [@holcomb_measuring_2025]*
+![Schematic representation of the GEDI data structure](figs/beam_product_footprint.png){#fig1}
+*Fig. 1: A schematic representation of the GEDI data structure. Credits: Amelia Holcomb's PhD dissertation [@holcomb_measuring_2025].*
 
 Several efforts in the broader NASA LiDAR community have tackled similar challenges. For example, SlideRule [@Shean2023] provides a scalable, cloud-based framework for processing ICESat-2 photon data, enabling users to query and transform complex satellite LiDAR datasets into analysis-ready forms. This illustrates a common pattern: while raw LiDAR missions deliver highly relevant observations, the data formats and scales make direct scientific use difficult without specialised infrastructure. For GEDI, existing services such as NASA's GEDI Subsetter via the Multi-Mission Algorithm and Analysis Platform (MAAP) [@chuck_daniels_2025_15122227] offer useful access for small to moderate-scale extractions, but they are not designed for operational-scale workflows or integration into reproducible pipelines. Moreover, MAAP introduces two important limitations for many researchers:  
 
@@ -52,7 +52,7 @@ Several efforts in the broader NASA LiDAR community have tackled similar challen
 The increasing use of GEDI in global applications, such as canopy height mapping [@pauls2024estimatingcanopyheightscale], disturbance assessment [@HOLCOMB2024114174], and forest degradation monitoring [@bourgoin2024human], highlights the need for efficient and scalable tooling. By streamlining data access and enabling large-scale, reproducible workflows, `gediDB` provides a scalable framework for GEDI data analysis, complementing efforts like SlideRule in the ICESat-2 domain and supporting broader ecological monitoring and policy-relevant research.
 
 
-![<a name="fig2"></a>Schematic representation of the gediDB workflow](figs/GEDIDB_FLOWCHART.png)
+![Schematic representation of the gediDB workflow](figs/GEDIDB_FLOWCHART.png){#fig2}
 *Fig. 2: A schematic representation of the gediDB data workflow.*
 
 
@@ -93,7 +93,7 @@ GEDI observations are inherently sparse: laser footprints sample only a tiny fra
 
 Comprehensive metadata, covering provenance, units, variable descriptions, and versioning, is embedded within the TileDB arrays, ensuring transparency and reproducibility.
 
-![<a name="fig3"></a>TileDB fragment schema for GEDI data](figs/tileDB_fragment_structure.png)
+![TileDB fragment schema for GEDI data](figs/tileDB_fragment_structure.png){#fig3}
 *Fig. 3: Illustration of the global GEDI data storage schema using TileDB arrays.*
 
 
@@ -126,7 +126,10 @@ These benchmarks demonstrate that `gediDB` consistently outperforms both MAAP an
 
 # Example use cases
 
[…]
