### 2025-09-25 d4de804
fix reference to figure
