### 2025-09-25 dc26bf5
fix bug in issue hyperlink
