### 2025-09-25 #8593 @jhkennedy
@kthyng I understand 250-1000 words is in the journal guidelines for papers, but I really don't see how the author could cut 1/2 the manuscript. Very few of the sections are longer than a brief description and this is in-line with the length of the Sliderule paper (https://github.com/openjournals/joss-reviews/issues/4982#issuecomment-1332569948), which is a similar piece of software. Do you have a suggestion on what should be cut while still including the expected items?
https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper
