### 2025-09-26 #8593 @kanishkan91
@simonbesnard1 - Would it be possible to move some text to the README? Right now everything from "PhD dissertation" to "Rich data integration" seems important but not for the paper. This can be moved to the documentation or even the README (I would prefer the documentation, but up to you). Even the "Future development" section belongs in the README or docs I think.

That should help cut this down.
