### 2025-09-26 dc11f63
remove future development from paper
