### 2025-09-26 ca0cd95
condense core functionality section
