### 2025-09-26 #8593 @simonbesnard1
@kanishkan91 I have made the paper a bit more concise by moving the  'future development' section to the README and making the 'core functionalities' section way shorter. I also shortened the statement of need section quite a bit. Let me know if those changes are sufficient.
